fix(policy): clean up + fixes

+++ api/db/migrations/policy/20191212100000_create_policy_trip_view.up.sql
@@ -4,7 +4,7 @@ CREATE MATERIALIZED VIEW policy.trips AS (
     cp.trip_id as trip_id,
     cp.start_insee as start_insee,
     cp.end_insee as end_insee,
-    cp.operator_id as operator_id,
+    cp.operator_id::int as operator_id,
     cp.operator_class as operator_class,
     cp.datetime as datetime, 
     cp.seats as seats,
